app/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "auth service"
authors = ["Deepayan Mukherjee <deepayanmukherjee1312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
motor = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
